+++ hack/runtime-migrator/cmd/backup/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/init"
+	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/initialisation"
 	"log/slog"
 	"os"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -12,9 +12,9 @@ import (
 
 func main() {
 	slog.Info("Starting runtime-backuper")
-	cfg := init.NewConfig()
+	cfg := initialisation.NewConfig()
 
-	init.PrintConfig(cfg)
+	initialisation.PrintConfig(cfg)
 
 	opts := zap.Options{
 		Development: true,
@@ -24,13 +24,13 @@ func main() {
 
 	gardenerNamespace := fmt.Sprintf("garden-%s", cfg.GardenerProjectName)
 
-	kubeconfigProvider, err := init.SetupKubernetesKubeconfigProvider(cfg.GardenerKubeconfigPath, gardenerNamespace, expirationTime)
+	kubeconfigProvider, err := initialisation.SetupKubernetesKubeconfigProvider(cfg.GardenerKubeconfigPath, gardenerNamespace, expirationTime)
 	if err != nil {
 		slog.Error(fmt.Sprintf("Failed to create kubeconfig provider: %v", err))
 		os.Exit(1)
 	}
 
-	shootClient, err := init.SetupGardenerShootClient(cfg.GardenerKubeconfigPath, gardenerNamespace)
+	shootClient, err := initialisation.SetupGardenerShootClient(cfg.GardenerKubeconfigPath, gardenerNamespace)
 	if err != nil {
 		slog.Error("Failed to setup Gardener shoot client", slog.Any("error", err))
 		os.Exit(1)
@@ -43,7 +43,7 @@ func main() {
 	}
 
 	slog.Info("Reading runtimeIds from input file")
-	runtimeIds, err := init.GetRuntimeIDsFromInputFile(cfg)
+	runtimeIds, err := initialisation.GetRuntimeIDsFromInputFile(cfg)
 	if err != nil {
 		slog.Error("Failed to read runtime Ids from input", slog.Any("error", err))
 		os.Exit(1)
