+++ hack/runtime-migrator/cmd/migration/migration.go
@@ -8,7 +8,7 @@ import (
 
 	gardener_types "github.com/gardener/gardener/pkg/client/core/clientset/versioned/typed/core/v1beta1"
 	runtimev1 "github.com/kyma-project/infrastructure-manager/api/v1"
-	config2 "github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/init"
+	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/initialisation"
 	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/migration"
 	"github.com/kyma-project/infrastructure-manager/pkg/config"
 	"github.com/kyma-project/infrastructure-manager/pkg/gardener/kubeconfig"
@@ -26,7 +26,7 @@ type Migration struct {
 	isDryRun        bool
 }
 
-func NewMigration(migratorConfig config2.Config, converterConfig config.ConverterConfig, auditLogConfig auditlogs.Configuration, kubeconfigProvider kubeconfig.Provider, kcpClient client.Client, shootClient gardener_types.ShootInterface) (Migration, error) {
+func NewMigration(migratorConfig initialisation.Config, converterConfig config.ConverterConfig, auditLogConfig auditlogs.Configuration, kubeconfigProvider kubeconfig.Provider, kcpClient client.Client, shootClient gardener_types.ShootInterface) (Migration, error) {
 
 	outputWriter, err := migration.NewOutputWriter(migratorConfig.OutputPath)
 	if err != nil {
@@ -44,7 +44,7 @@ func NewMigration(migratorConfig config2.Config, converterConfig config.Converte
 }
 
 func (m Migration) Do(ctx context.Context, runtimeIDs []string) error {
-	listCtx, cancel := context.WithTimeout(ctx, config2.TimeoutK8sOperation)
+	listCtx, cancel := context.WithTimeout(ctx, initialisation.TimeoutK8sOperation)
 	defer cancel()
 
 	shootList, err := m.shootClient.List(listCtx, v1.ListOptions{})
@@ -92,7 +92,7 @@ func (m Migration) Do(ctx context.Context, runtimeIDs []string) error {
 			return
 		}
 
-		migrationCtx, cancel := context.WithTimeout(ctx, config2.TimeoutK8sOperation)
+		migrationCtx, cancel := context.WithTimeout(ctx, initialisation.TimeoutK8sOperation)
 		defer cancel()
 
 		runtime, err := m.runtimeMigrator.Do(migrationCtx, *shootToMigrate)
