+++ hack/runtime-migrator/cmd/restore/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	init2 "github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/init"
+	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/initialisation"
 	"log/slog"
 	"os"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -11,9 +11,9 @@ import (
 
 func main() {
 	slog.Info("Starting runtime-restorer")
-	cfg := init2.NewRestoreConfig()
+	cfg := initialisation.NewRestoreConfig()
 
-	init2.PrintRestoreConfig(cfg)
+	initialisation.PrintRestoreConfig(cfg)
 
 	opts := zap.Options{
 		Development: true,
@@ -23,19 +23,19 @@ func main() {
 
 	gardenerNamespace := fmt.Sprintf("garden-%s", cfg.GardenerProjectName)
 
-	_, err := init2.SetupKubernetesKubeconfigProvider(cfg.GardenerKubeconfigPath, gardenerNamespace, expirationTime)
+	_, err := initialisation.SetupKubernetesKubeconfigProvider(cfg.GardenerKubeconfigPath, gardenerNamespace, expirationTime)
 	if err != nil {
 		slog.Error(fmt.Sprintf("Failed to create kubeconfig provider: %v", err))
 		os.Exit(1)
 	}
 
-	_, err = init2.CreateKcpClient(&cfg.Config)
+	_, err = initialisation.CreateKcpClient(&cfg.Config)
 	if err != nil {
 		slog.Error("Failed to create kcp client", slog.Any("error", err))
 		os.Exit(1)
 	}
 
-	_, err = init2.SetupGardenerShootClient(cfg.GardenerKubeconfigPath, gardenerNamespace)
+	_, err = initialisation.SetupGardenerShootClient(cfg.GardenerKubeconfigPath, gardenerNamespace)
 	if err != nil {
 		slog.Error("Failed to setup Gardener shoot client", slog.Any("error", err))
 		os.Exit(1)
