+++ hack/runtime-migrator/internal/initialisation/params.go
@@ -1,4 +1,4 @@
-package init
+package initialisation
 
 import (
 	"bufio"
