+++ hack/runtime-migrator/internal/restore/restorer.go
@@ -1,12 +1,12 @@
 package restore
 
 import (
-	init2 "github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/init"
+	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/initialisation"
 	"github.com/kyma-project/infrastructure-manager/pkg/gardener/kubeconfig"
 )
 
 type Restorer struct {
-	cfg                init2.Config
+	cfg                initialisation.Config
 	isDryRun           bool
 	kubeconfigProvider kubeconfig.Provider
 }
